Rethinking of DataEngine
